Extract HelixTool.Mcp.Tools, add Core NuGet packaging (W1-W6, W9)
- W6: Centralize version (0.2.1) in Directory.Build.props, remove from HelixTool.csproj
- W1: Create HelixTool.Mcp.Tools class library, move HelixMcpTools.cs + McpToolResults.cs from Core
- W2: Update project references — CLI, Mcp, and Tests now reference Mcp.Tools; remove ModelContextProtocol from Core
- W4: Promote IsFileSearchDisabled from internal to public (needed by Mcp.Tools after split)
- W3: Extract 11 nested record types from HelixService to individual files in Core/Models/
- W5: Add NuGet packaging metadata to Core (PackageId: lewing.helix.core, MIT license, doc generation)
- W9: CI publish workflow packs both lewing.helix.mcp and lewing.helix.core; version validation reads Directory.Build.props

Note: src/HelixTool/.mcp/server.json still has hardcoded version fields that must match — no change to the release process.
