+++ .github/workflows/publish.yml
@@ -30,12 +30,12 @@ jobs:
       - name: Validate version consistency
         run: |
           TAG_VERSION="${{ steps.tag_name.outputs.current_version }}"
-          CSPROJ_VERSION=$(grep -oP '(?<=<Version>)[^<]+' src/HelixTool/HelixTool.csproj)
+          CSPROJ_VERSION=$(grep -oP '(?<=<Version>)[^<]+' Directory.Build.props)
           SERVER_VERSION=$(jq -r '.version' src/HelixTool/.mcp/server.json)
           PACKAGE_VERSION=$(jq -r '.packages[0].version' src/HelixTool/.mcp/server.json)
           ERRORS=""
           if [ "$CSPROJ_VERSION" != "$TAG_VERSION" ]; then
-            ERRORS="${ERRORS}HelixTool.csproj Version: expected '${TAG_VERSION}', got '${CSPROJ_VERSION}'\n"
+            ERRORS="${ERRORS}Directory.Build.props Version: expected '${TAG_VERSION}', got '${CSPROJ_VERSION}'\n"
           fi
           if [ "$SERVER_VERSION" != "$TAG_VERSION" ]; then
             ERRORS="${ERRORS}server.json top-level version: expected '${TAG_VERSION}', got '${SERVER_VERSION}'\n"
@@ -50,13 +50,16 @@ jobs:
           fi
           echo "All versions match tag: ${TAG_VERSION}"
 
-      - name: Pack
-        run: dotnet pack src/HelixTool -c Release -o src/HelixTool/nupkg /p:Version=${{ steps.tag_name.outputs.current_version }}
+      - name: Pack CLI tool
+        run: dotnet pack src/HelixTool -c Release -o nupkg /p:Version=${{ steps.tag_name.outputs.current_version }}
+
+      - name: Pack Core library
+        run: dotnet pack src/HelixTool.Core -c Release -o nupkg /p:Version=${{ steps.tag_name.outputs.current_version }}
 
       - name: Create Release
         uses: ncipollo/release-action@v1
         with:
-          artifacts: src/HelixTool/nupkg/*.nupkg
+          artifacts: nupkg/*.nupkg
           token: ${{ secrets.GITHUB_TOKEN }}
           body: "Release ${{ github.ref_name }}"
 
@@ -67,4 +70,4 @@ jobs:
           user: ${{ secrets.NUGET_USER }}
 
       - name: Push to NuGet
-        run: dotnet nuget push "src/HelixTool/nupkg/*.nupkg" --source https://api.nuget.org/v3/index.json --api-key ${{ steps.login.outputs.NUGET_API_KEY }}
+        run: dotnet nuget push "nupkg/*.nupkg" --source https://api.nuget.org/v3/index.json --api-key ${{ steps.login.outputs.NUGET_API_KEY }}
