+++ Directory.Build.props
@@ -1,0 +1,7 @@
+<Project>
+
+  <PropertyGroup>
+    <Version>0.2.1</Version>
+  </PropertyGroup>
+
+</Project>
