+++ src/HelixTool.Core/HelixTool.Core.csproj
@@ -6,10 +6,19 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <PropertyGroup Label="Package Metadata">
+    <PackageId>lewing.helix.core</PackageId>
+    <Description>.NET client library for the Helix test infrastructure API</Description>
+    <Authors>Larry Ewing</Authors>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageTags>helix;dotnet;ci;testing;api-client</PackageTags>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <PublishRepositoryUrl>true</PublishRepositoryUrl>
+  </PropertyGroup>
+
   <ItemGroup>
     <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.7" />
     <PackageReference Include="Microsoft.DotNet.Helix.Client" Version="11.0.0-beta.26110.116" />
-    <PackageReference Include="ModelContextProtocol" Version="1.0.0" />
   </ItemGroup>
 
   <ItemGroup>
