+++ src/HelixTool.Core/Models/BatchJobSummary.cs
@@ -1,0 +1,4 @@
+namespace HelixTool.Core;
+
+/// <summary>Summary for multiple jobs.</summary>
+public record BatchJobSummary(List<JobSummary> Jobs, int TotalFailed, int TotalPassed);
