+++ src/HelixTool.Core/Models/FileEntry.cs
@@ -1,0 +1,4 @@
+namespace HelixTool.Core;
+
+/// <summary>Represents an uploaded file from a Helix work item.</summary>
+public record FileEntry(string Name, string Uri);
