+++ src/HelixTool.Core/Models/LogSearchResult.cs
@@ -1,0 +1,4 @@
+namespace HelixTool.Core;
+
+/// <summary>Result of searching a console log.</summary>
+public record LogSearchResult(string WorkItem, List<LogMatch> Matches, int TotalLines);
