+++ src/HelixTool.Mcp.Tools/HelixMcpTools.cs
