+++ src/HelixTool.Mcp.Tools/McpToolResults.cs
