+++ src/HelixTool.Tests/HelixTool.Tests.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\HelixTool.Core\HelixTool.Core.csproj" />
+    <ProjectReference Include="..\HelixTool.Mcp.Tools\HelixTool.Mcp.Tools.csproj" />
     <ProjectReference Include="..\HelixTool.Mcp\HelixTool.Mcp.csproj" />
   </ItemGroup>
